pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tilecraft"
version = "0.1.0"
description = "GPU kernel practice codebase for Triton and NVIDIA CuTe DSL."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.26",
]

[project.optional-dependencies]
triton = [
    "triton>=3.0; platform_system == 'Linux'",
]
cute-cu12 = [
    "nvidia-cutlass-dsl>=4.4; platform_system == 'Linux'",
]
cute-cu13 = [
    "nvidia-cutlass-dsl[cu13]>=4.4; platform_system == 'Linux'",
]
torch = [
    "torch>=2.4; platform_system == 'Linux'",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.5",
]

[tool.setuptools]
packages = ["cute_dsl_kernels", "triton_kernels"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
